+++ src/store/base/types.js
@@ -1,6 +1,9 @@
 // Base types
 
 // Successfully loaded header image
+export const LOAD_PLACEHOLDER_SUCCESS = 'LOAD_PLACEHOLDER_SUCCESS';
+
+// Successfully loaded header image
 export const LOAD_HEADER_IMAGE_SUCCESS = 'LOAD_HEADER_IMAGE_SUCCESS';
 
 // Successfully loaded about page sections