+++ changelog.md
@@ -7,19 +7,50 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.0.3
 
+### 2018-11-15
+
+#### Added
+
+- Sorting between past and future projects
+- Future project item layout and styling
+- Current project sorting
+- Project page layout and styling
+- Retreived Lilac image placeholder from [Contentful](https://www.contentful.com/)
+  - Added new Lilac image placeholder [Redux](https://redux.js.org) thunk
+- Added project page layout and styling
+- Added global page item component
+- Minor styling
+
+#### Removed
+
+- Global project item naming
+- `LOAD_HEADER_IMAGE_SUCCESS` type
+- Replaced blog item component contents with blog item
+
+### 2018-11-14
+
+#### Removed
+
+- Changed all intances of `work` to `project` for projects feature
+
 ### 2018-11-13
 
 #### Added
+
 - About page layout and styling
 - About item layout and styling
 - Retreived about items from [Contentful](https://www.contentful.com/)
 
 #### Removed
+
 - Unecessary sorting functions
 
 ### 2018-11-12
 
+s
+
 #### Added
+
 - Added figures scss module
   - Added zoom image hover and grayscale hover
 - Added no select styling
@@ -31,12 +62,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Blog list and grid hover
 
 #### Removed
+
 - Renamed blog item component
 - Blog grid hover overlay
 
 ### 2018-11-08
 
 #### Added
+
 - `TEAM_LOADING` Team thunk type
 - Team member profile page routing
 - Redid team member profile page layout and styling
@@ -47,6 +80,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Team member image hover and grayscale
 
 #### Removed
+
 - `LOAD_TEAM_MEMBERS_SUCCESS` Team thunk type
 - Team page introduction blurb
 - Role in team member items on team member page
@@ -56,11 +90,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 2018-11-07
 
 #### Added
- - Able to retreive header images from [Contentful](https://www.contentful.com/)
-   - Implemented through new base [Redux](https://redux.js.org)
+
+- Able to retreive header images from [Contentful](https://www.contentful.com/)
+  - Implemented through new base [Redux](https://redux.js.org)
 - Navbar toggler placement and colouring
 - Retreived Lilac Foundation social media from [Contentful](https://www.contentful.com/)
-  - Added new socials [Redux](https://redux.js.org) reducer
+  - Added new socials [Redux](https://redux.js.org) thunk
 - Linked footer links to respective pages
 - Separated [Contentful](https://www.contentful.com/) into `blogClient` and `contentClient`
 - Coloured links mixin improvements
@@ -68,6 +103,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 2018-11-06
 
 #### Added
+
 - Page header layout, styling, and responsivness
 - Global overlay styling
 - Coloured links styling
@@ -76,6 +112,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 2018-11-02
 
 #### Added
+
 - Poly-fluid sizing typography courtesy of [Jakobud](https://github.com/Jakobud)
 - Colour modules
 - New navbar layout
@@ -85,6 +122,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Footer content, layout, and styling
 
 #### Removed
+
 - Component specific typography partials
 - Navbar styling
   - Background colour
@@ -96,30 +134,34 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 2018-10-31
 
 #### Added
+
 - Work post scene
 - Work post routing
 - Work page layout
   - Retreived work page content from [Contentful](https://www.contentful.com/)
-   - Work page work post items
+  - Work page work post items
 - Markdown content image styling
 
 #### Removed
+
 - Blog content
 
 ### 2018-10-30
 
 #### Added
+
 - Team member route
 - Team member profile page
 - Global team member component
   - Team member component layout
 - Global page header component
-- New blog functions path 
+- New blog functions path
 - Added work scene
   - Added work route
 - Retreived work posts from [Contentful](https://www.contentful.com/)
 
 #### Removed
+
 - About page team section item
 - Responsive layout styling
 - Blog page header component
@@ -129,16 +171,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 2018-10-26, 2018-10-29
 
 #### Added
+
 - Added blog page header
 - Implemented blog card layout
 - Blog post content layout
 
 #### Removed
+
 - Blog item link props
 
 ### 2018-10-25
 
 #### Added
+
 - [Fontawesome](https://fontawesome.com/) icons
 - Updated footer layout, styling, and typography
 - Added Scroll To Top route component
@@ -151,6 +196,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 2018-10-24
 
 #### Added
+
 - Sorted blog post by descending date
 - Updated home blog section layout, styling, and responsiveness
 - Page section height prop
@@ -165,6 +211,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Simplified about section layout
 
 #### Removed
+
 - Blog section item
 - Team section folder mispelling
 - Loader class
@@ -172,6 +219,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 2018-10-22
 
 #### Added
+
 - Added blog item layout and card styling
 - Added Work item layout and styling
 - Added Team member item layout and styling
@@ -184,6 +232,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 2018-10-21
 
 #### Added
+
 - Implemented home page layout with styling
   - Added Mission section
   - Added Team section
@@ -194,6 +243,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 2018-10-10, 2018-10-17, 2018-10-18, 2018-10-20
 
 #### Added
+
 - Retreived & displayed blog posts from [Contentful](https://www.contentful.com/)
 - Retreived & displayed single blog post from [Contentful](https://www.contentful.com/) with blog post id
 - Basic styling with [Bootstrap](https://getbootstrap.com)
@@ -208,6 +258,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 2018-10-09
 
 #### Added
+
 - Iniital commit
 - [Webpack](https://webpack.js.org/) configuration
 - [Babel](https://babeljs.io/) configuraiton